Parse NH committees from dynamic data.
